fix verifier to load a program if an exit is present at any place
With -O2, clang might decide to place the exit of a code that always
ends in the middle of the code. Anyway, the check of the exit as the
last opcode is clearly insufficient, as one can create eBPF bytecode
that always loops before calling the last exit instruction. This fix
allows possible eligible code to run, although the verifier should
actually perform a DFS with jump instructions to check if the exit
instruction is potentially reachable.